node: incorporate search into modified files structure
Rearrange prior search commit in a few files and parameters to fit
modified structure of the app.

Fix JS style errors.

Seems to give an error on the "home page" search...
